Move Error into submodule
