+++ crates/module-system/sov-modules-api/src/gas/meters/mod.rs
@@ -1,6 +1,6 @@
 mod basic;
 mod slot;
 mod unlimited;
-pub use basic::BasicGasMeter;
+pub use basic::{BasicGasMeter, GasInfo};
 pub use slot::SlotGasMeter;
 pub use unlimited::UnlimitedGasMeter;
