+++ crates/module-system/sov-modules-api/src/gas/mod.rs
@@ -1,3 +1,4 @@
+mod error;
 mod impl_macros;
 mod metered_utils;
 mod meters;
@@ -8,6 +9,7 @@ mod unit;
 #[cfg(test)]
 mod tests;
 
+pub use error::GasMeteringError;
 pub(crate) use impl_macros::*;
 pub use metered_utils::{
     charge_gas_to_deserialize_json, metered_credential, MeteredBorshDeserialize,
